+++ packages/assert/mod.ts
@@ -1,3 +1,3 @@
-export * from "./assert_ok.ts";
-export * from "./assert_header.ts";
-export * from "./assert_status.ts";
+export * from "./ok.ts";
+export * from "./header.ts";
+export * from "./status.ts";